Create directories and move files
