+++ README.md
@@ -1,1 +1,5 @@
-# CrazyCart_Systemanzeige
+[![Version: 1.0 Release](https://img.shields.io/badge/Version-1.0%20Release-green.svg)](https://github.com/Sprudlwassr) [![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
+
+# Systemanzeige
+
+Kurzbeschreibung
